Now deletes member profile upon membership termination, can only revoke bank authorization if no open orderlines.

+++ app/Http/Controllers/BankController.php
@@ -140,6 +140,10 @@ class BankController extends Controller
             Session::flash("flash_message", "As a member you cannot revoke your bank authorization. You can update it, though.");
             return Redirect::back();
         }
+        if ($user->hasUnpaidOrderlines()) {
+            Session::flash("flash_message", "You cannot revoke your bank authorization while you still have unpaid orderlines.");
+            return Redirect::back();
+        }
         $user->bank->delete();
 
         Session::flash("flash_message", "Deleted bank account.");
